disable cc when debug=trye
