+++ gradle-plugin-testing-junit/src/main/java/com/palantir/gradle/testing/execution/DefaultGradleInvoker.java
@@ -19,7 +19,6 @@ package com.palantir.gradle.testing.execution;
 import com.google.common.collect.ImmutableList;
 import com.google.errorprone.annotations.RestrictedApi;
 import com.palantir.gradle.testing.RestrictedCreation;
-import java.lang.management.ManagementFactory;
 import java.nio.file.Path;
 import java.util.Arrays;
 import org.gradle.testkit.runner.GradleRunner;
@@ -32,7 +31,7 @@ public record DefaultGradleInvoker(Path rootProjectDir, GradleVersion gradleVers
     public GradleInvocation withArgs(String... args) {
         GradleRunner runner = GradleRunner.create()
                 .withProjectDir(rootProjectDir.toFile())
-                .withDebug(shouldRunInTestkitDebugMode())
+                .withDebug(GradleInvoker.shouldRunInTestkitDebugMode())
                 .forwardOutput()
                 .withGradleVersion(gradleVersion.version())
                 .withPluginClasspath()
@@ -42,27 +41,5 @@ public record DefaultGradleInvoker(Path rootProjectDir, GradleVersion gradleVers
                         .build());
 
         return new DefaultGradleInvocation(runner);
-    }
-
-    private static boolean shouldRunInTestkitDebugMode() {
-        // `withDebug(true)` will run the Gradle daemon inside the same JVM as the test, whereas
-        // `withDebug(false)` will run Gradle in a new daemon.
-        // When running tests from IntelliJ with debug or coverage, they only work when the Gradle daemon
-        // is in the same the JVM as the test, so we must set `withDebug(true)` in these cases.
-        // Beware: There can be differences between these two modes!
-        return isJavaDebugAgentLoaded() || isRunningCoverageInIntelliJ();
-    }
-
-    private static boolean isJavaDebugAgentLoaded() {
-        // When you run a test with debug in intellij, it passes an arg to the test process like:
-        //   -agentlib:jdwp=transport=dt_socket,server=n,suspend=y,address=127.0.0.1:54342
-        return ManagementFactory.getRuntimeMXBean().getInputArguments().stream()
-                .anyMatch(arg -> arg.contains("-agentlib:jdwp"));
-    }
-
-    private static boolean isRunningCoverageInIntelliJ() {
-        // When you run a test with coverage in intellij, it sets a system property on the test JVM
-        // by adding the jvm arg `-Didea.coverage.calculate.hits=true`.
-        return Boolean.getBoolean("idea.coverage.calculate.hits");
     }
 }
